Move test to `other` testsuite
